Move manual chargeback tests from test_manual.py and parametrize them.
