+++ cfme/tests/test_manual.py
@@ -1,4 +1,3 @@
-# pylint: skip-file
 """Manual tests"""
 import pytest
 
@@ -79,248 +78,6 @@ def test_sui_stack_service_vm_detail_page_should_show_correct_data():
         casecomponent: SelfServiceUI
         caseimportance: medium
         initialEstimate: 1/4h
-    """
-    pass
-
-
-@pytest.mark.manual
-def test_validate_chargeback_cost_weekly_rate_network_cost():
-    """
-    Validate network I/O used cost in a daily Chargeback report by
-    assigning weekly rate
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-def test_validate_chargeback_cost_monthly_rate_disk_cost():
-    """
-    Validate disk I/O used cost in a daily Chargeback report by assigning
-    monthly rate
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-@test_requirements.chargeback
-def test_validate_chargeback_cost_monthlyreport_hourly_rate_network_cost():
-    """
-    Validate network I/O used cost in a monthly Chargeback report
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-@test_requirements.chargeback
-def test_validate_chargeback_cost_weeklyreport_hourly_rate_disk_cost():
-    """
-    Validate disk I/O used cost in a weekly Chargeback report
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-def test_validate_chargeback_cost_monthly_rate_network_cost():
-    """
-    Validate network I/O used cost in a daily Chargeback report by
-    assigning monthly rate
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-def test_validate_chargeback_cost_weekly_rate_disk_cost():
-    """
-    Validate disk I/O used cost in a dailyy Chargeback report by assigning
-    weekly rate
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-@test_requirements.chargeback
-def test_validate_chargeback_cost_weekly_rate_cpu_cost():
-    """
-    Validate CPU usage cost in a daily Chargeback report by assigning
-    weekly rate
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-@test_requirements.chargeback
-def test_validate_chargeback_cost_monthlyreport_hourly_rate_disk_cost():
-    """
-    Validate disk I/O used cost in a monthly Chargeback report
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-@test_requirements.chargeback
-def test_validate_chargeback_cost_weeklyreport_hourly_rate_cpu_cost():
-    """
-    Validate CPU usage cost in a weekly Chargeback report
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-@test_requirements.chargeback
-def test_validate_chargeback_cost_monthlyreport_hourly_rate_memory_cost():
-    """
-    Validate memory usage cost in a monthly Chargeback report
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-def test_validate_chargeback_cost_monthly_rate_cpu_cost():
-    """
-    Validate CPU usage cost in a daily Chargeback report by assigning
-    monthly rate
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-@test_requirements.chargeback
-def test_validate_chargeback_cost_weekly_rate_memory_cost():
-    """
-    Validate memory usage cost in a daily Chargeback report by assigning
-    weekly rate
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-@test_requirements.chargeback
-def test_validate_chargeback_cost_weeklyreport_hourly_rate_memory_cost():
-    """
-    Validate memory usage cost in a weekly Chargeback report
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-@test_requirements.chargeback
-def test_validate_chargeback_cost_weeklyreport_hourly_rate_network_cost():
-    """
-    Validate network I/O used cost in a weekly Chargeback report
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-@test_requirements.chargeback
-def test_validate_chargeback_cost_monthlyreport_hourly_rate_cpu_cost():
-    """
-    Validate CPU usage cost in a monthly Chargeback report
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-def test_validate_chargeback_cost_monthly_rate_memory_cost():
-    """
-    Validate memory usage cost in a daily Chargeback report by assigning
-    monthly rate
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
     """
     pass
 
@@ -472,57 +229,6 @@ def test_heat_stacks_in_non_admin_tenants_shall_also_be_collected():
 
 
 @pytest.mark.manual
-@test_requirements.chargeback
-@pytest.mark.tier(2)
-def test_chargeback_resource_allocation_cpu_allocated():
-    """
-    Verify CPU allocation in a Chargeback report based on resource
-    allocation. C&U data is not considered for these reports.
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-@test_requirements.chargeback
-@pytest.mark.tier(2)
-def test_chargeback_resource_allocation_memory_allocated():
-    """
-    Verify memory allocation in a Chargeback report based on resource
-    allocation.C&U data is not considered for these reports.
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
-@test_requirements.chargeback
-@pytest.mark.tier(2)
-def test_chargeback_resource_allocation_storage_allocated():
-    """
-    Verify storage allocation in a Chargeback report based on resource
-    allocation. C&U data is not considered for these reports.
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: medium
-        initialEstimate: 1/10h
-    """
-    pass
-
-
-@pytest.mark.manual
 @test_requirements.service
 @pytest.mark.tier(2)
 def test_bundle_stack_deployment():
@@ -665,22 +371,6 @@ def test_button_groups_created_on_orchestration_type_heat_service_catalog_items_
         casecomponent: Services
         caseimportance: low
         initialEstimate: 1/4h
-    """
-    pass
-
-
-@pytest.mark.manual
-@test_requirements.chargeback
-@pytest.mark.tier(3)
-def test_saved_chargeback_report_show_full_screen():
-    """
-    Verify that saved chargeback reports can be viewed
-
-    Polarion:
-        assignee: tpapaioa
-        casecomponent: CandU
-        caseimportance: low
-        initialEstimate: 1/12h
     """
     pass
